+++ src/resource/index.ts
@@ -20,6 +20,7 @@
 //
 
 export { getFile, overwriteFile, saveFileInContainer } from "./file";
+export { createContainerInContainer } from "./createContainerInContainer";
 export { getSolidDataset } from "./getSolidDataset";
 export { saveSolidDatasetAt } from "./saveSolidDatasetAt";
 export { saveSolidDatasetInContainer } from "./saveSolidDatasetInContainer";